+++ runtime/service/src/main/java/org/apache/polaris/service/config/ServiceProducers.java
@@ -69,6 +69,8 @@ import org.apache.polaris.service.auth.external.OidcConfiguration;
 import org.apache.polaris.service.auth.external.tenant.OidcTenantResolver;
 import org.apache.polaris.service.auth.internal.broker.TokenBroker;
 import org.apache.polaris.service.auth.internal.broker.TokenBrokerFactory;
+import org.apache.polaris.service.authz.AuthorizationConfiguration;
+import org.apache.polaris.service.authz.AuthorizationRealmConfiguration;
 import org.apache.polaris.service.catalog.api.IcebergRestOAuth2ApiService;
 import org.apache.polaris.service.catalog.io.FileIOConfiguration;
 import org.apache.polaris.service.catalog.io.FileIOFactory;
@@ -157,19 +159,20 @@ public class ServiceProducers {
   }
 
   @Produces
-  @ApplicationScoped
-  public PolarisAuthorizerFactory polarisAuthorizerFactory(
-      AuthorizationConfiguration authorizationConfig,
-      @Any Instance<PolarisAuthorizerFactory> authorizerFactories) {
-    PolarisAuthorizerFactory factory =
-        authorizerFactories.select(Identifier.Literal.of(authorizationConfig.type())).get();
-    return factory;
+  @RequestScoped
+  public AuthorizationRealmConfiguration realmAuthzConfig(
+      AuthorizationConfiguration config, RealmContext realmContext) {
+    return config.forRealm(realmContext);
   }
 
   @Produces
   @RequestScoped
   public PolarisAuthorizer polarisAuthorizer(
-      PolarisAuthorizerFactory factory, RealmConfig realmConfig) {
+      AuthorizationRealmConfiguration authorizationConfig,
+      @Any Instance<PolarisAuthorizerFactory> factories,
+      RealmConfig realmConfig) {
+    PolarisAuthorizerFactory factory =
+        factories.select(Identifier.Literal.of(authorizationConfig.type())).get();
     return factory.create(realmConfig);
   }